change logic of obj_from_urn
originally we decided which object model to use based on the URI/URL
This is no longer possible (arguably not good practise either).  So
we now look for the SQLAlchemy model in each table that has the given
UID.  If we find it, we return the object.  We cannot know in advance
which model the UUID applies to (ie GET folder or GET collection) so
we try all three.  SqlAlchemy fights my attempts to search then retrieve.